Many bug fixes
Fixed PlaylistDetailScreen had dual titles.
Add more fallback APIs for Artist and Lyrics Data when RapidAPI will be down.
Fixed LRCLib API didn't show lyrics for songs with double data.
Fixed critical bug causing app to crash when navigating between screens causing java.lang.IndexOutOfBoundsException.

+++ app/src/main/java/chromahub/rhythm/app/network/LRCLibApiService.kt
@@ -31,7 +31,7 @@ data class LrcLibLyrics(
     @SerializedName("trackName") val trackName: String?,
     @SerializedName("artistName") val artistName: String?,
     @SerializedName("albumName") val albumName: String?,
-    @SerializedName("duration") val duration: Int?,
+    @SerializedName("duration") val duration: Double?,
     @SerializedName("instrumental") val instrumental: Boolean?,
     @SerializedName("plainLyrics") val plainLyrics: String?,
     @SerializedName("syncedLyrics") val syncedLyrics: String?
